fix: remove `always()` on commit step
there is a scenario where semantic-release can fail and subsequent runs
of this composite action won't fix the state of the repo due to the
merge commit already happening. specifically, when semantic-release
fails in between having created a commit and git tag, but before it
creates the github release. this causes semantic-release to not create a
release because the source branch is ahead of main.